refactoring typeayout, goodslist, goodscard components
