+++ src/App.css
@@ -3,10 +3,6 @@
   --main-yellow: #fce200;
   --text-grey: rgba(255, 255, 255, 0.7);
 }
-/* .App {
-  text-align: center;
-} */
-
 main,
 html {
   background-color: var(--background-main);
