+++ src/components/cosmetics/goods/goodsCard.module.css
@@ -1,0 +1,21 @@
+.grid {
+  row-gap: 0.6rem;
+  /* use Grid container and Grid items from mui - response change items */
+}
+.cardActionArea {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  height: 100%;
+}
+.card {
+  height: 100%;
+  max-width: 100%;
+  margin: 0.3rem;
+}
+.cardContent {
+  color: var(--text-grey);
+  font-weight: 500;
+  text-align: center;
+  /* font size change in GoodsCard*/
+}
